+++ lib/views/auth_views/reset_password_page.dart
@@ -4,7 +4,10 @@
 /// TODO: edit this page
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
 import 'package:xiaokeai/components/main_view.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:xiaokeai/shared/views/ui_consts.dart';
 
 class ResetPasswordPage extends StatefulWidget {
   const ResetPasswordPage({super.key});
@@ -21,71 +24,110 @@ class _ResetPasswordPageState extends State<ResetPasswordPage> {
   Widget build(BuildContext context) {
     return mainView(
       context,
+      appBarTitle: AppLocalizations.of(context)!.resetPassword.toUpperCase(),
       appBarBackgroundColor: Colors.transparent,
       extendBodyBehindAppBar: true,
       body: Container(
         decoration: BoxDecoration(
           gradient: LinearGradient(
-            begin: Alignment.topRight,
-            end: Alignment.bottomLeft,
-            colors: [Colors.green[400]!, Colors.teal[500]!],
+            end: Alignment.topCenter,
+            begin: Alignment.bottomCenter,
+            colors: [
+              Theme.of(context).colorScheme.primaryContainer,
+              Theme.of(context).colorScheme.secondaryContainer,
+              Theme.of(context).colorScheme.tertiaryContainer
+            ],
           ),
         ),
-        child: SafeArea(
-          child: Center(
-            child: SingleChildScrollView(
+        child: Center(
+          child: SingleChildScrollView(
+            child: Container(
+              constraints: BoxConstraints(maxWidth: 550),
+              margin: UiConsts.PaddingAll_large,
+              transformAlignment: AlignmentDirectional.centerStart,
               child: Padding(
-                padding: const EdgeInsets.all(30.0),
+                padding: UiConsts.PaddingAll_large,
                 child: Form(
                   key: _formKey,
                   child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
+                    mainAxisAlignment: MainAxisAlignment.spaceAround,
+                    crossAxisAlignment: CrossAxisAlignment.stretch,
                     children: [
                       Icon(
                         Icons.lock_reset,
-                        size: 100,
-                        color: Colors.white,
+                        size: 110,
+                        color: Theme.of(context).colorScheme.onPrimary,
                       ),
                       SizedBox(height: 30),
                       Text(
                         'Enter your email to reset your password',
-                        style: TextStyle(color: Colors.white, fontSize: 18),
+                        style: TextStyle(
+                            color: Theme.of(context).colorScheme.onSurface,
+                            fontSize: 18),
                         textAlign: TextAlign.center,
                       ),
-                      SizedBox(height: 30),
+                      UiConsts.SizedBoxGapVertical_large,
                       TextFormField(
-                        controller: _emailController,
-                        decoration: InputDecoration(
-                          hintText: 'Email',
-                          filled: true,
-                          fillColor: Colors.white.withOpacity(0.8),
-                          border: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(30),
-                            borderSide: BorderSide.none,
-                          ),
-                        ),
-                        validator: (value) {
-                          if (value == null || value.isEmpty) {
-                            return 'Please enter your email';
-                          }
-                          return null;
-                        },
-                      ),
-                      SizedBox(height: 30),
+                          controller: _emailController,
+                          autofocus: false,
+                          validator: FormBuilderValidators.email(),
+                          autovalidateMode: AutovalidateMode.onUserInteraction,
+                          decoration: InputDecoration(
+                              prefixIcon: Icon(Icons.email_rounded),
+                              hintText: AppLocalizations.of(context)!.email,
+                              labelText: AppLocalizations.of(context)!.email,
+                              floatingLabelBehavior: FloatingLabelBehavior.auto,
+                              floatingLabelAlignment:
+                                  FloatingLabelAlignment.center,
+                              errorStyle: TextStyle(
+                                  color: Theme.of(context)
+                                      .colorScheme
+                                      .inverseSurface),
+                              errorBorder: UnderlineInputBorder(
+                                  borderRadius:
+                                      UiConsts.BorderRadiusCircular_standard,
+                                  borderSide: BorderSide(
+                                    color: Theme.of(context).colorScheme.error,
+                                  )),
+                              focusedErrorBorder: UnderlineInputBorder(
+                                borderRadius:
+                                    UiConsts.BorderRadiusCircular_standard,
+                                borderSide: BorderSide(
+                                    color: Theme.of(context).colorScheme.error),
+                              ),
+                              focusedBorder: OutlineInputBorder(
+                                borderRadius:
+                                    UiConsts.BorderRadiusCircular_standard,
+                                borderSide: BorderSide(
+                                    color: Theme.of(context).focusColor),
+                              ),
+                              enabledBorder: OutlineInputBorder(
+                                borderRadius:
+                                    UiConsts.BorderRadiusCircular_standard,
+                                borderSide: BorderSide(
+                                    color: Theme.of(context)
+                                        .colorScheme
+                                        .onPrimary),
+                              ),
+                              border: InputBorder.none)),
+                      UiConsts.SizedBoxGapVertical_large,
                       ElevatedButton(
                         onPressed: _resetPassword,
                         style: ElevatedButton.styleFrom(
                           foregroundColor:
-                              Theme.of(context).colorScheme.onPrimaryContainer,
+                              Theme.of(context).colorScheme.onTertiaryContainer,
                           backgroundColor:
-                              Theme.of(context).colorScheme.primaryContainer,
+                              Theme.of(context).colorScheme.tertiaryContainer,
                           padding: EdgeInsets.symmetric(
                               horizontal: 50, vertical: 15),
                           shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(30),
+                            borderRadius:
+                                UiConsts.BorderRadiusCircular_standard,
                           ),
                         ),
-                        child: Text('Reset Password'),
+                        child: Text(AppLocalizations.of(context)!
+                            .resetPassword
+                            .toUpperCase()),
                       ),
                     ],
                   ),